Add sale-detail calculated-fields and update nullable-fields to not null in db and ef model.
Adds sale and sale-detail total.
Updates sale-detail CurrentPrice, Quantity and Discount to not null.
Deletes unnecessary convertions in sales controller and service.

+++ SalesSystem.Business/Services/SalesService.cs
@@ -79,7 +79,7 @@ namespace SalesSystem.Business.Services
                 }
 
                 // Update product stock
-                product.Stock -= (decimal)saleDetail.Quantity;
+                product.Stock -= saleDetail.Quantity;
 
                 saleDetail.CurrentPrice = product.Price;
             }
